pyproject.toml
[tool.poetry]
name = "ocr-pipeline"
version = "0.1.0"
description = ""
authors = ["Yaiphaba-Ng <yaiphabalamzing@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
evaluate = "^0.4.3"
datasets = "^3.0.1"
jiwer = "^3.0.4"
accelerate = "^0.34.2"
pyarrow = "^17.0.0"
scikit-learn = "^1.5.2"
torch = "^2.4.1"
torchvision = "^0.19.1"
tqdm = "^4.66.5"
matplotlib = "^3.9.2"
transformers = {extras = ["torch"], version = "^4.45.1"}
opencv-python = "^4.10.0.84"
python-doctr = {extras = ["torch"], version = "^0.9.0"}
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
python-multipart = "^0.0.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
